+++ react/FilterNavigator.js
@@ -76,6 +76,8 @@ const FilterNavigator = ({
   scrollToTop = 'none',
   openFiltersMode = 'many',
   filtersFetchMore,
+  fullWidthOnMobile = false,
+  navigationTypeOnMobile = 'page',
 }) => {
   const { isMobile } = useDevice()
   const handles = useCssHandles(CSS_HANDLES)
@@ -189,6 +191,9 @@ const FilterNavigator = ({
               priceRange={priceRange}
               preventRouteChange={preventRouteChange}
               navigateToFacet={navigateToFacet}
+              fullWidth={fullWidthOnMobile}
+              navigationType={navigationTypeOnMobile}
+              initiallyCollapsed={initiallyCollapsed}
             />
           </div>
         </div>