+++ react/components/AccordionFilterGroup.js
@@ -20,6 +20,8 @@ const AccordionFilterGroup = ({
   open,
   onOpen,
   onFilterCheck,
+  navigationType,
+  initiallyCollapsed,
 }) => {
   const handles = useCssHandles(CSS_HANDLES)
   const filters = useSelectedFilters(facets)
@@ -35,6 +37,8 @@ const AccordionFilterGroup = ({
       show={show}
       onOpen={onOpen}
       quantitySelected={quantitySelected}
+      navigationType={navigationType}
+      initiallyCollapsed={initiallyCollapsed}
     >
       <div
         className={classNames(